requirements.txt
sqlalchemy
streamlit 
pyodbc
python-dotenv
pandas
langchain
langchain-community 
langchain-core
langchain-google-genai